requirements.txt
appdirs==1.4.4
flask_cors==4.0.0
discord==2.3.2